+++ main.py
@@ -1,21 +1,22 @@
 # Package: BulkDNS
 # Module: main
 # Author: Michal Selma <michal@selma.cc>
-# Rev: 2023-11-25
+# Rev: 2023-12-10
 
 import os
 import sys
 import configparser
+import logging
 
 from init import init_core
+
 from common import logger
 
-log = logger.log_run()
-
+log = logging.getLogger('main')
 
 if __name__ == '__main__':
-    # init logger
-    logger.log_configure(log)
+    # init logging configuration for root logger
+    logger.configure(log)
 
     # Read configuration
     config = configparser.ConfigParser()
@@ -23,4 +24,16 @@ if __name__ == '__main__':
         log.critical(f'Configuration file not found.')
         sys.exit(0)
 
-    init_core.run(config)
+    log.info('1 - System initialization')
+    log.info('2 - Domains check')
+    log.info('3 - Archiving')
+    log.info('Choose option and press Enter: ')
+    user_option = input()
+    if user_option == '1':
+        init_core.run(config)
+    elif user_option == '2':
+        log.info('Not implemented yet...')
+    elif user_option == '3':
+        log.info('Not implemented yet...')
+    else:
+        log.info('Incorrect option picked')